Entry moved to Model namespace
